add privats.h (read there), correct time_ntp, renamed and mor stability

+++ sensor/ESP8266/time_ntp.h
@@ -31,7 +31,7 @@ date_time_t;
 unsigned long getNTPTimestamp(IPAddress& ipaddr);
 unsigned long sendNTPpacket(IPAddress& address);
 unsigned int date_time_to_epoch(date_time_t* date_time);
-void epoch_to_date_time(date_time_t* date_time,unsigned int epoch);
-String epoch_to_string(unsigned int epoch);
+void epoch_to_date_time(date_time_t* date_time,unsigned long epoch);
+String epoch_to_string(unsigned long epoch);
 
 #endif
